test leptonica

+++ util/run_stage.sh
@@ -238,10 +238,14 @@ fi
 PRESERVE_DLL_PATTERN="${DLL_PRESERVE_LIST:-openvino|torch|tensorflow|vulkan|amf|nvcodec|mfx|onevpl}"
 if [[ ! "$STAGENAME" =~ $PRESERVE_DLL_PATTERN ]]; then
     if [[ -d "$FFBUILD_DESTDIR$FFBUILD_PREFIX" ]]; then
-        log_info "################################################################"
-        log_debug "${BROOM_MARK} Cleaning unwanted DLLs from static stage: $STAGENAME"
-        # find "$FFBUILD_DESTDIR$FFBUILD_PREFIX" -type f -name "*.dll" -delete || true
-        find "$FFBUILD_DESTDIR$FFBUILD_PREFIX" -type f \( -name "*.dll" -o -name "*.dll.a" \) -delete || true
+        local DELETED_FILES=$(find "$FFBUILD_DESTDIR$FFBUILD_PREFIX" -type f \( -name "*.dll" -o -name "*.dll.a" \) -print)
+        if [[ -n "$DELETED_FILES" ]]; then
+            log_info "################################################################"
+            log_debug "${BROOM_MARK} Cleaning unwanted dynamic DLLs from static stage: $STAGENAME"
+            log_info "$DELETED_FILES" | sed "s|$FFBUILD_DESTDIR||g"
+            # find "$FFBUILD_DESTDIR$FFBUILD_PREFIX" -type f -name "*.dll" -delete || true
+            find "$FFBUILD_DESTDIR$FFBUILD_PREFIX" -type f \( -name "*.dll" -o -name "*.dll.a" \) -delete || true
+        fi
     else
         log_debug "${DIRS_MARK} No standard prefix directory to clean for $STAGENAME"
     fi
